move es index validation script
