+++ seqr/management/commands/lift_project_to_hg38.py
@@ -6,12 +6,12 @@ from pyliftover.liftover import LiftOver
 
 from reference_data.models import GENOME_VERSION_GRCh38
 from seqr.models import Project, SavedVariant, Sample
-from seqr.views.utils.dataset_utils import match_and_update_search_samples, \
-    validate_index_metadata_and_get_elasticsearch_index_samples
+from seqr.views.utils.dataset_utils import match_and_update_search_samples
 from seqr.views.utils.json_to_orm_utils import update_model_from_json
 from seqr.views.utils.orm_to_json_utils import get_json_for_saved_variants
 from seqr.views.utils.variant_utils import reset_cached_search_results
-from seqr.utils.search.utils import get_es_variants_for_variant_tuples, get_single_es_variant
+from seqr.utils.search.utils import get_es_variants_for_variant_tuples, get_single_es_variant, \
+    validate_index_metadata_and_get_elasticsearch_index_samples
 from seqr.utils.xpos_utils import get_xpos
 
 logger = logging.getLogger(__name__)
